requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
celery==5.3.4
django-otp==1.2.4